Modularize autologging utilities (#4216)
* rename back to autologging_utils


* Partial


* Add files


* Fixes


* More


* Explanatory comment


* Test fixes


* Format


* Remove raise


* Import fix


* Test


* Lint


* Comment


* Modularize versions, format


* Add versioning


* Remove import
