+++ mlflow/tracking/fluent.py
@@ -18,7 +18,7 @@ from mlflow.tracking.context import registry as context_registry
 from mlflow.store.tracking import SEARCH_MAX_RESULTS_DEFAULT
 from mlflow.utils import env
 from mlflow.utils.autologging_utils import (
-    _is_testing,
+    is_testing,
     autologging_integration,
     AUTOLOGGING_INTEGRATIONS,
     autologging_is_disabled,
@@ -1404,7 +1404,7 @@ def autolog(
             ) and not autologging_params.get("silent", False):
                 _logger.info("Autologging successfully enabled for %s.", module.__name__)
         except Exception as e:
-            if _is_testing():
+            if is_testing():
                 # Raise unexpected exceptions in test mode in order to detect
                 # errors within dependent autologging integrations
                 raise
@@ -1433,7 +1433,7 @@ def autolog(
         if "pyspark" in str(ie):
             register_post_import_hook(setup_autologging, "pyspark", overwrite=True)
     except Exception as e:
-        if _is_testing():
+        if is_testing():
             # Raise unexpected exceptions in test mode in order to detect
             # errors within dependent autologging integrations
             raise
