+++ tests/autologging/test_autologging_utils.py
@@ -12,22 +12,22 @@ import mlflow
 from mlflow.utils import gorilla
 from mlflow.tracking.client import MlflowClient
 from mlflow.utils.autologging_utils import (
+    AUTOLOGGING_INTEGRATIONS,
     log_fn_args_as_params,
-    _wrap_patch,
     resolve_input_example_and_signature,
     batch_metrics_logger,
     AutologgingEventLogger,
-    AutologgingSession,
     BatchMetricsLogger,
     autologging_integration,
     get_autologging_config,
     autologging_is_disabled,
-    _is_autologging_integration_supported,
-    _check_version_in_range,
-    _cross_tested_flavor_to_module_name_and_module_key,
 )
-from mlflow.utils.autologging_utils import AUTOLOGGING_INTEGRATIONS
-
+from mlflow.utils.autologging_utils.safety import _wrap_patch, AutologgingSession
+from mlflow.utils.autologging_utils.versioning import (
+    FLAVOR_TO_MODULE_NAME_AND_VERSION_INFO_KEY,
+    _check_version_in_range,
+    is_flavor_supported_for_associated_package_versions,
+)
 
 from tests.autologging.fixtures import test_mode_off
 
@@ -745,16 +745,18 @@ _module_version_info_dict_patch = {
     ],
 )
 @mock.patch(
-    "mlflow.utils.autologging_utils._module_version_info_dict", _module_version_info_dict_patch
+    "mlflow.utils.autologging_utils.versioning._module_version_info_dict",
+    _module_version_info_dict_patch,
 )
 def test_is_autologging_integration_supported(flavor, module_version, expected_result):
-    module_name, _ = _cross_tested_flavor_to_module_name_and_module_key[flavor]
+    module_name, _ = FLAVOR_TO_MODULE_NAME_AND_VERSION_INFO_KEY[flavor]
     with mock.patch(module_name + ".__version__", module_version):
-        assert expected_result == _is_autologging_integration_supported(flavor)
+        assert expected_result == is_flavor_supported_for_associated_package_versions(flavor)
 
 
 @mock.patch(
-    "mlflow.utils.autologging_utils._module_version_info_dict", _module_version_info_dict_patch
+    "mlflow.utils.autologging_utils.versioning._module_version_info_dict",
+    _module_version_info_dict_patch,
 )
 def test_disable_for_unsupported_versions_warning_sklearn_integration():
     log_warn_fn_name = "mlflow.utils.autologging_utils._logger.warning"
